// frontend/src/types/TVSeriesDetail.d.ts
type TVSeriesDetail = {
    adult: boolean;
    backdrop_path: string | null;
    created_by: Array<{
        id: number;
        credit_id: string;
        name: string;
        gender: number;
        profile_path: string | null;
    }>;
    episode_run_time: Array<number>;
    first_air_date: string;
    genres: Array<{ id: number; name: string }>;
    homepage: string;
    id: number;
    in_production: boolean;
    languages: Array<string>;
    last_air_date: string;
    last_episode_to_air: {
        air_date: string;
        episode_number: number;
        id: number;
        name: string;
        overview: string;
        production_code: string;
        season_number: number;
        still_path: string | null;
        vote_average: number;
        vote_count: number;
    };
    name: string;
    next_episode_to_air: null;
    networks: Array<{
        name: string;
        id: number;
        logo_path: string | null;
        origin_country: string;
    }>;
    number_of_episodes: number;
    number_of_seasons: number;
    origin_country: Array<string>;
    original_language: string;
    original_name: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    production_companies: Array<{
        id: number;
        logo_path: string | null;
        name: string;
        origin_country: string;
    }>;
    production_countries: Array<{ iso_3166_1: string; name: string }>;
    seasons: Array<{
        air_date: string;
        episode_count: number;
        id: number;
        name: string;
        overview: string;
        poster_path: string | null;
        season_number: number;
        vote_average: number;
        vote_count: number;
    }>;
    spoken_languages: Array<{ english_name: string; iso_639_1: string; name: string }>;
    status: string;
    tagline: string;
    type: string;
    vote_average: number;
    vote_count: number;
};

export default TVSeriesDetail;